alter column price and balance to type money add constraint email

+++ tables/3_items.sql
@@ -3,7 +3,7 @@ CREATE TABLE items(
     uid UUID NOT NULL DEFAULT uuid_generate_v4(),
     name TEXT NOT NULL CHECK(length(name) <= 200),
     description TEXT DEFAULT '' CHECK(length(description) <= 500),
-    price NUMERIC(10,2) NOT NULL CHECK(price > 0),
+    price MONEY NOT NULL CHECK(price > MONEY(0)),
     seller_pid INTEGER NOT NULL REFERENCES sellers(pid),
     amount INTEGER NOT NULL CHECK(amount >= 0)
 );
